+++ package.json
@@ -34,14 +34,15 @@
     }
   },
   "peerDependencies": {
-    "@inertiajs/vue3": "1.0.0",
-    "vue": "^3.x"
+    "@inertiajs/vue3": "^1.0.0",
+    "vue": "^3.x",
+    "axios": "^1.2.0"
   },
   "devDependencies": {
     "@types/node": "^18.0.0",
-    "axios": "^0.21.1",
+    "axios": "^1.2.0",
     "typescript": "^4.5.4",
-    "vite": "^3.2.2",
-    "vue-tsc": "^0.38.1"
+    "vite": "^4.0.4",
+    "vue-tsc": "^1.0.24"
   }
 }
