+++ src/use-modal.ts
@@ -4,17 +4,7 @@ import { defineAsyncComponent, h, nextTick, watch, computed, ref, shallowRef } f
 import axios from "axios"
 import resolver from "./resolver"
 
-interface Modal {
-  component: string
-  baseURL: string
-  redirectURL: string | null
-  props: Record<string, any>
-  key: string
-  nonce: string
-}
-
-const response = usePage<{ modal: Modal }>().props
-const modal = computed(() => response.value?.modal)
+const modal = computed(() => usePage().props?.modal)
 const props = computed(() => modal.value?.props)
 const key = computed(() => modal.value?.key)
 
@@ -94,7 +84,7 @@ if (typeof window !== "undefined") {
 }
 
 watch(
-  () => modal.value,
+  modal,
   () => {
     if (modal.value?.nonce !== nonce.value) {
       resolveComponent()
@@ -102,7 +92,8 @@ watch(
   },
   { deep: true }
 )
-watch(() => key.value, updateHeaders)
+
+watch(key, updateHeaders)
 
 const redirect = () => {
   var redirectURL = modal.value?.redirectURL ?? modal.value?.baseURL
